# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Online Examination Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attempts:
    get:
      operationId: listAttempts
      tags: [attempts]
      summary: List exam attempts
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: examId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
    post:
      operationId: startAttempt
      tags: [attempts]
      summary: Start exam attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptInput"
      responses:
        "201":
          description: Attempt started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetail"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Get attempt with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptDetail"
  /attempts/{id}/answer:
    post:
      operationId: submitAnswer
      tags: [attempts]
      summary: Submit answer for a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer saved
  /attempts/{id}/submit:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit exam (finish attempt)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # RESULTS
...
  /analytics/exams/{examId}:
    get:
      operationId: getExamAnalytics
      tags: [analytics]
      summary: Exam-level analytics
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamAnalytics"
  /analytics/students/{studentId}:
    get:
      operationId: getStudentAnalytics
      tags: [analytics]
      summary: Student performance analytics
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAnalytics"
  /analytics/lecturer/{lecturerId}:
    get:
      operationId: getLecturerAnalytics
      tags: [analytics]
      summary: Lecturer dashboard analytics
      parameters:
        - name: lecturerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecturer analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LecturerAnalytics"
components:
  schemas:
...
    Attempt:
      type: object
      required: [id, studentId, examId, status, startedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        examId:
          type: integer
        examTitle:
          type: ["string", "null"]
        status:
          type: string
          enum: [in_progress, submitted, graded]
        score:
          type: ["number", "null"]
        startedAt:
          type: string
        submittedAt:
          type: ["string", "null"]
    AttemptInput:
      type: object
      required: [examId]
      properties:
        examId:
          type: integer
    AttemptDetail:
      type: object
      required: [id, studentId, examId, status, startedAt, questions]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        examId:
          type: integer
        examTitle:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        status:
          type: string
        startedAt:
          type: string
        submittedAt:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AttemptQuestion"
    AttemptQuestion:
      type: object
      required: [id, questionId, text, type, points]
      properties:
        id:
          type: integer
        questionId:
          type: integer
        text:
          type: string
        type:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        points:
          type: number
        studentAnswer:
          type: ["string", "null"]
        codeLanguage:
          type: ["string", "null"]
    AnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string

    # RESULT
    Result:
      type: object
      required: [id, attemptId, studentId, examId, score, totalPoints, isPassed, createdAt]
      properties:
        id:
          type: integer
        attemptId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        examId:
          type: integer
        examTitle:
          type: ["string", "null"]
        score:
          type: number
        totalPoints:
          type: number
        percentage:
          type: number
        isPassed:
          type: boolean
        gradingStatus:
          type: string
          enum: [auto_graded, pending_manual, fully_graded]
        createdAt:
          type: string
...
    ExamAnalytics:
      type: object
      required: [examId, examTitle, totalAttempts, averageScore, passRate, highestScore, lowestScore]
      properties:
        examId:
          type: integer
        examTitle:
          type: string
        totalAttempts:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number
        highestScore:
          type: number
        lowestScore:
          type: number
        questionPerformance:
          type: array
          items:
            $ref: "#/components/schemas/QuestionStat"
        scoreDistribution:
          type: array
          items:
            $ref: "#/components/schemas/CountByLabel"
    StudentAnalytics:
      type: object
      required: [studentId, studentName, totalAttempts, averageScore, passRate]
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        totalAttempts:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        coursePerformance:
          type: array
          items:
            $ref: "#/components/schemas/CourseStat"
    LecturerAnalytics:
      type: object
      required: [lecturerId, lecturerName, totalCourses, totalExams, totalStudents]
      properties:
        lecturerId:
          type: integer
        lecturerName:
          type: string
        totalCourses:
          type: integer
        totalExams:
          type: integer
        totalStudents:
          type: integer
        averageClassScore:
          type: number
        examSummaries:
          type: array
          items:
            $ref: "#/components/schemas/ExamSummary"
...
    CountByLabel:
      type: object
      required: [label, count]
      properties:
        label:
          type: string
        count:
          type: number
    QuestionStat:
      type: object
      required: [questionId, questionText, correctRate]
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        correctRate:
          type: number
        totalAnswers:
          type: integer
    CourseStat:
      type: object
      required: [courseId, courseName, averageScore]
      properties:
        courseId:
          type: integer
        courseName:
          type: string
        averageScore:
          type: number
        examCount:
          type: integer
    ExamSummary:
      type: object
      required: [examId, examTitle, attemptCount, averageScore]
      properties:
        examId:
          type: integer
        examTitle:
          type: string
        attemptCount:
          type: integer
        averageScore:
          type: number
        passRate:
          type: number