# lib/api-zod/src/generated/types/courseStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Online Examination Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseStat {
  courseId: number;
  courseName: string;
  averageScore: number;
  examCount?: number;
}